+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * `Holdings - Call number, eye readable`. Add search only by the call number. Refs UIIN-1040.
 * Replace `bigtest/mirage` with `miragejs`.
 * `Item - Effective call number (item), eye readable`. Add search only by the call number. Refs UIIN-1041.
+* Move (right to left) and confirm items and holdings or items to a new instance. Refs UIIN-1152.
 
 ## [4.0.1](https://github.com/folio-org/ui-inventory/tree/v4.0.1) (2020-07-01)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v4.0.0...v4.0.1)